requirements.txt
# 核心依赖 - 更新到最新稳定版本
PyYAML==6.0.1
click==8.1.7
APScheduler==3.10.4
requests==2.31.0
python-dotenv==1.0.0
argcomplete==3.2.2

# AI服务相关
anthropic==0.18.0
openai==1.12.0

# Git操作相关
GitPython==3.1.40
PyGithub==2.1.1
python-gitlab==4.2.0

# 文档处理
python-docx==1.1.0
pandas==2.2.0
python-pptx==0.6.23

# 机器学习相关
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2

# 图像处理
Pillow==10.2.0
opencv-python==4.9.0.80

# PDF处理
PyPDF2==3.0.1

# 日志和工具
colorlog==6.8.2
python-dateutil==2.8.2
networkx==3.2.1

# 安全扫描工具
safety==3.0.1
bandit==1.7.5

# 开发和测试
pytest==8.0.0
pytest-mock==3.12.0
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# 性能监控
psutil==5.9.8
memory-profiler==0.61.0
